Replace deprecated features in unrelated runnable tests
This fixes several deprecation warnings for older tests concerning:
- D1 style operator overloads
- foreach index type coercion
- Integer promotion
- Initializing immutable data from `static this()`

+++ test/runnable/test42.d
@@ -1972,7 +1972,7 @@ struct Foobar;
 int test124()
 {   int result;
     dchar[] aa;
-    alias uint foo_t;
+    alias size_t foo_t;
 
     foreach (foo_t i, dchar d; aa)
     {
